download_sra_reads.sh: Created. It downloads SRA runs from SRA experiments or directly. User can provide the layout of all the SRA IDs or they will be inferred from ID metadata. Metadata is retrieved by chunks and sleep time. The chunk size and sleep time can be determined by the user. OMAdataDownloader.sh:
Avoided double double-quoted strings (breaks the string) on line 165.
Avoided creating a subshell on line 350; instead, used curly braces {} to group commands.
Removed unnecessary cat on line 419.

Step2.sh:

On line 489, I had forgotten to use double quotes in the echo command!

In both scripts:

Added ./. The ./ prefix specifies that the file is located in the current directory. This is especially useful to prevent filenames starting with hyphens from being misinterpreted as command options.
Removed the leading portion of the script's pathname using basename for the usage function.